Add files via upload
Set modem on right side of turtles. This is a working version of the monster miner. The 1x1Tunnel program goes on all of miningTurtles. The sendCommands goes onto the commandTurtle. Then the followingTurtle program goes with followingTurtle. Listener is just extra code that I was using but implemented into the 1x1Tunnel.